regexai/requirements.txt
Django==5.2.1
python-decouple==3.8
gunicorn==22.0.0
whitenoise==6.9.0
groq==0.29.0
razorpay==1.4.2
